move tests to their right directory finally
